bot/requirements.txt
python-telegram-bot==20.7
vosk==0.3.45
pydub==0.25.1
python-dotenv==1.0.0
psutil==5.9.5
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.97
protobuf>=3.20.0
accelerate>=0.20.0
ffmpeg-python==0.2.0
numpy==1.24.3
scipy==1.11.4
noisereduce==3.0.3
fastapi==0.104.1
uvicorn==0.24.0
pyttsx3==2.90
schedule==1.2.0
pywin32==306